tests/submission/PhilipTaylor/canvas/2d.state.saverestore.stack.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.state.saverestore.stack</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.state.html">state</a>.<a href="index.2d.state.saverestore.html">saverestore</a>.stack</h1>
<p class="desc">save()/restore() can be nested as a stack</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.state.save">2d.state.save</a>
</li><li><a href="../annotated-spec/canvas.html#testrefs.2d.state.restore">2d.state.restore</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("save()/restore() can be nested as a stack");
_addTest(function(canvas, ctx) {

ctx.lineWidth = 1;
ctx.save();
ctx.lineWidth = 2;
ctx.save();
ctx.lineWidth = 3;
_assertEqual(ctx.lineWidth, 3, "ctx.lineWidth", "3");
ctx.restore();
_assertEqual(ctx.lineWidth, 2, "ctx.lineWidth", "2");
ctx.restore();
_assertEqual(ctx.lineWidth, 1, "ctx.lineWidth", "1");


});
</script>

<div id="log"></div>
</body></html>
